test: replace MainServiceTest with BaseServiceTest
